+++ examples/basic/cut_freehand.py
@@ -55,6 +55,7 @@ vedo.settings.use_parallel_projection = True  # to avoid perspective artifacts
 
 msh = vedo.Volume(vedo.dataurl+'embryo.tif').isosurface().color('gold', 0.25) # Mesh
 
-plt = FreeHandCutPlotter(msh).add_hover_legend()
+plt = FreeHandCutPlotter(msh)
+plt.add_hover_legend()
 #plt.init(some_list_of_initial_pts) #optional!
 plt.start(axes=1, bg2='lightblue').close()